+++ admin/assets/functions/functions-admin.php
@@ -1,10 +1,7 @@
 <?php
 
-// Imports
-
-include('env.php');
-
 // BDD
+require('env.php');
 try {
     $bdd = new PDO('mysql:host=localhost;dbname=tagalog;charset=utf8', $user, $mdp);
 }
@@ -12,4 +9,7 @@ catch (Exception $e) {
     die('Errer : ' . $e->getMessage());
 }
 
+// Select form - Articles from "travels"
+
+// Queries
 require('queries-admin.php');