+++ admin/assets/manageTravels/category-travels.php
@@ -7,6 +7,7 @@
         <table >
             <thead>
                 <tr>
+                    <th>Preview picture</th>
                     <th>Category</th>
                     <th>Update</th>
                     <th>Delete</th>
@@ -15,8 +16,9 @@
             <tbody>
                 <?php while ($donnees = $categories_travel_select->fetch()) { ?>
                     <tr>
+                        <td><img src="../assets/pictures/travels/categories/<?php echo $donnees['thumbnail']; ?>"></td>
                         <td><?php echo $donnees['category']; ?></td>
-                        <td><a href="update-form-admin.php?update=travels&display=category&idArticle=<?php echo $donnees['id']; ?>"><i class="fas fa-edit"></i></a></td>
+                        <td><a href="update-form-admin.php?update=travels&display=category&id=<?php echo $donnees['id']; ?>"><i class="fas fa-edit"></i></a></td>
                         <td><a href="assets/manageTravels/controller/delete.php?manageTravels=category&id=<?php echo $donnees['id']; ?>"><i class="fas fa-trash-alt"></i></a></td>
                     </tr>
                 <?php } ?>
@@ -27,7 +29,10 @@
         <h4>Add a category</h4>
         <form action="assets/manageTravels/controller/add.php?manageTravels=category" method="POST">
             <div>
-                <input type="text" name="thumbnail" placeholder="Picture.ext">
+                <label class="btn-file-input-preview">
+                    <i class="fa fa-image"></i><span class="span-no-choice">Choose category picture</span><span class="span-file-chosen"></span><input type="file" style="display: none;" name="thumbnail">
+                </label>
+                <button type="button" class="btn-delete-picture"><i class="far fa-times-circle"></i></button>
             </div>
             <div>
                 <input type="text" name="category-list" placeholder="Category">