+++ admin/assets/manageTravels/controller/add.php
@@ -6,6 +6,7 @@ var_dump($_POST);
 
 if (isset($_POST)) {
     if (isset($_GET['manageTravels']) && $_GET['manageTravels'] == 'content') {
+        
         if (isset($_POST['title']) && $_POST['title'] != '') {
             $addDate = date('d/m/Y');
             $picture = $_POST['picture'];
@@ -35,41 +36,40 @@ if (isset($_POST)) {
                 'display_carousel' => $display_carousel,
                 'category_id' => $category,
             ))) {
-                $confirmMessage = 'The article '.$title.' is well added in the database. Thanks for your add ! :D';
+                $confirmMessage = 'The article '.$title.' is well added in the database. Thanks for your add !  <i class="far fa-laugh-beam"></i>';
                 header('location: ../../../confirm-portal-admin.php?manageTravels=content&message='.$confirmMessage.'');
             }
             else {
-                $errorMessage = 'There\'s a problem while entering the entry in the database <i class="fas fa-frown-open"></i>';
+                $errorMessage = 'The article '.$title.' wasn\'t saved - Error while joining the database (or with queries)  <i class="far fa-tired"></i>';
                 header('location: ../../../error-portal-admin.php?manageTravels=content&message='.$errorMessage.'');
             }
         }
-    }
-    else {
-        $errorMessage = 'You must fill a category and picture input at least <i class="fas fa-frown-open"></i>';
-        header('location: ../../../error-portal-admin.php?manageTravels=content&message='.$errorMessage.'');
-    }
 
-    if (isset($_GET['manageTravels']) && $_GET['manageTravels'] == 'category') {
-        if (isset($_POST['category-list']) && $_POST['category-list'] != '') {
+    }
+    else if (isset($_GET['manageTravels']) && $_GET['manageTravels'] == 'category') {
+        
+        if (isset($_POST['category-list']) && $_POST['category-list'] != '' && isset($_POST['thumbnail']) && $_POST['thumbnail'] != '') {
             $category = $_POST['category-list'];
             $thumbnail = $_POST['thumbnail'];
         
-            $add = $bdd->prepare('INSERT INTO categories_travel (category) VALUES (:category)');
+            $add = $bdd->prepare('INSERT INTO categories_travel (category, thumbnail) VALUES (:category, :thumbnail)');
         
             if ($add->execute(array(
-                'category' => $category
+                'category' => $category,
+                'thumbnail' => $thumbnail
             ))) {
-                $confirmMessage = 'The category '.$category.' is well added in the database. Thanks for your add ! :D';
+                $confirmMessage = 'The category '.$category.' is well added in the database. Thanks for your add !  <i class="far fa-laugh-beam"></i>';
                 header('location: ../../../confirm-portal-admin.php?manageTravels=category&message='.$confirmMessage.'');
             }
             else {
-                $errorMessage = 'The category '.$category.' wasn\'t saved - Error while joining the database (or with queries) :/';
+                $errorMessage = 'The category '.$category.' wasn\'t saved - Error while joining the database (or with queries)  <i class="far fa-tired"></i>';
                 header('location: ../../../error-portal-admin.php?manageTravels=category&message='.$errorMessage.'');
             }
         }
         else {
-            $errorMessage = 'You must fill a category :/';
+            $errorMessage = 'You must fill a category and a picture here  <i class="far fa-smile"></i>';
             header('location: ../../../error-portal-admin.php?manageTravels=category&message='.$errorMessage.'');
         }
+
     }
 }