+++ admin/assets/manageTravels/controller/update.php
@@ -6,13 +6,23 @@ var_dump($_POST);
 
 if (isset($_POST['thumbnail'])) {
 
-    if ($_POST['thumbnail'] != '' && isset($_POST['category-list']) && $_POST['category-list'] != '') {
+    if (isset($_POST['category-list']) && $_POST['category-list'] != '') {
+        $id = $_GET['id'];
+        // $thumbnail = $_POST['thumbnail'];
 
-        $id = $_GET['idArticle'];
-        $thumbnail = $_POST['thumbnail'];
+        if ($_POST['thumbnail'] != '') {
+            $thumbnail = $_POST['thumbnail'];  
+        }
+        else {
+            $preview_picture = $bdd->query('SELECT thumbnail FROM categories_travel WHERE id = "'.$id.'"');
+            while ($data = $preview_picture->fetch()) {
+                $thumbnail = $data['thumbnail'];
+            }
+        }
+
         $category = $_POST['category-list'];
     
-        $update = $bdd->prepare('UPDATE categories_travel SET thumbnail = "'.$thumbnail.'", category = "'.$category.'" WHERE id = '.$id.'');
+        $update = $bdd->query('UPDATE categories_travel SET thumbnail = "'.$thumbnail.'", category = "'.$category.'" WHERE id = '.$id.'');
     
         if ($update->execute()) {
             $confirmMessage = 'The article '.$_POST['category-list'].' is now updated in the database !';