+++ admin/assets/manageTravels/update-travels.php
@@ -10,11 +10,12 @@
             <div>
                 <select name="category">
                     <option value="0">---</option>
-                    <?php while ($donnees = $categories_travel_select->fetch()) { ?>
-                        <option value="<?php echo $donnees['id']; ?>"><?php echo $donnees['category']; ?></option>
+                    <?php while ($select = $categories_travel_select->fetch()) { ?>
+                        <option value="<?php echo $select['id']; ?>"><?php echo $select['category']; ?></option>
                     <?php } ?>
                 </select>
             </div>
+            <?php while($donnees = $update_travels_content->fetch()) { ?>
             <div>
                 <label for="file" class="label-file-input-preview">Choose preview picture</label>
                 <input id="file" type="file" name="picture" style="display: none;">
@@ -41,6 +42,7 @@
                     </div>
                     <hr>
                 </div>
+                <?php } ?>
             </div>
             <div>
                 <input type="text" name="title" placeholder="Title">
@@ -68,10 +70,16 @@
     <article class="form-update">
         <?php while($donnees = $update_travels_category->fetch()) { ?>
         <h4>Update <?php echo $donnees['category']; ?> category</h4>
-        <form action="assets/manageTravels/controller/update.php?manageTravels=category&idArticle=<?php echo $donnees['id']; ?>" method="POST">
+        <form action="assets/manageTravels/controller/update.php?manageTravels=category&id=<?php echo $donnees['id']; ?>" method="POST">
             <div class="display-pictures-update">
+                <input type="hidden" name="path_pictures" value="../assets/pictures/travels/categories/">
                 <img src="../assets/pictures/travels/categories/<?php echo $donnees['thumbnail']; ?>" alt="thumbnail of <?php echo $donnees['category']; ?> - travels category">
-                <input type="text" name="thumbnail" placeholder="Picture.ext" value="<?php echo $donnees['thumbnail']; ?>">
+                <div class="d-flex justify-content-center align-items-center">
+                    <label class="btn-file-input-preview">
+                        <i class="fa fa-image mr-2"></i><span class="span-no-choice"><?php echo $donnees['thumbnail']; ?></span><span class="span-file-chosen"></span><input type="file" style="display: none;" name="thumbnail">
+                    </label>
+                    <button type="button" class="btn-delete-picture"><i class="far fa-times-circle"></i></button>
+                </div>
             </div>
             <div>
                 <input type="text" name="category-list" placeholder="Category" value="<?php echo $donnees['category']; ?>">
@@ -84,5 +92,4 @@
         <hr>
         <a href="manageTravels.php?manageTravels=category">Previous</a>
     </article>
-
-<?php } ?>
+    <?php } ?>