+++ assets/functions/functions.php
@@ -54,3 +54,7 @@ $nbday= $interval->format('%m');
 //     var_dump($donnees);
 // }
 
+// Display articles in different pages - Travel
+
+
+